🧹 Tidy: consolidate transcript storage logic into HandlesTranscriptStorage trait
