+++ app/Services/AudioTranscriptionService.php
@@ -8,6 +8,7 @@ use App\Contracts\TranscriptionServiceInterface;
 use App\Exceptions\NonRetryableTranscriptionException;
 use App\Exceptions\TranscriptionException;
 use App\Traits\DetectsStorageType;
+use App\Traits\HandlesTranscriptStorage;
 use Exception;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
@@ -18,6 +19,7 @@ use OpenAI\Laravel\Facades\OpenAI;
 class AudioTranscriptionService implements TranscriptionServiceInterface
 {
     use DetectsStorageType;
+    use HandlesTranscriptStorage;
 
     public function __construct(
         private readonly SermonProcessingLogger $logger,
@@ -448,74 +450,6 @@ class AudioTranscriptionService implements TranscriptionServiceInterface
         ];
 
         return in_array($exception->getStatusCode(), $nonRetryableCodes);
-    }
-
-    /**
-     * Store transcript content (delegates to TranscriptStorageService)
-     *
-     * @param  int  $sermonId  The sermon ID
-     * @param  string  $transcript  The transcript content
-     * @return string The file path where transcript was stored
-     *
-     * @throws Exception When storage fails
-     */
-    public function storeTranscript(int $sermonId, string $transcript): string
-    {
-        return $this->storageService->storeTranscript($sermonId, $transcript);
-    }
-
-    /**
-     * Retrieve transcript content (delegates to TranscriptStorageService)
-     *
-     * @param  int  $sermonId  The sermon ID
-     * @return string|null The transcript content or null if not found
-     */
-    public function getTranscript(int $sermonId): ?string
-    {
-        return $this->storageService->getTranscript($sermonId);
-    }
-
-    /**
-     * Check if transcript exists (delegates to TranscriptStorageService)
-     *
-     * @param  int  $sermonId  The sermon ID
-     * @return bool True if transcript exists
-     */
-    public function transcriptExists(int $sermonId): bool
-    {
-        return $this->storageService->transcriptExists($sermonId);
-    }
-
-    /**
-     * Delete transcript file (delegates to TranscriptStorageService)
-     *
-     * @param  int  $sermonId  The sermon ID
-     * @return bool True if deleted or didn't exist
-     */
-    public function deleteTranscript(int $sermonId): bool
-    {
-        return $this->storageService->deleteTranscript($sermonId);
-    }
-
-    /**
-     * Clean up transcript files on failure (delegates to TranscriptStorageService)
-     *
-     * @param  int  $sermonId  The sermon ID
-     */
-    public function cleanupOnFailure(int $sermonId): void
-    {
-        $this->storageService->cleanupOnFailure($sermonId);
-    }
-
-    /**
-     * Get the full transcript file path (delegates to TranscriptStorageService)
-     *
-     * @param  int  $sermonId  The sermon ID
-     * @return string The full file path
-     */
-    public function getTranscriptPath(int $sermonId): string
-    {
-        return $this->storageService->getTranscriptPath($sermonId);
     }
 
     /**
